Eve Online Market Forecaster/requirements.txt
about-time==3.1.1
alive-progress==2.4.1
asttokens==2.0.5
backcall==0.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fonttools==4.33.3
grapheme==0.6.0
idna==3.3
ipykernel==6.13.0
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.1
jupyter-core==4.10.0
kiwisolver==1.4.2
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==23.0.0
requests==2.27.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
stack-data==0.2.0
statsmodels==0.13.2
tornado==6.1
traitlets==5.2.1.post0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
